examples: one more timer example
Try to measure one TIM using another one. This example helps to understand
how much we can do in TIM interrupts. For instance, if TIM interrupt is
called too often in comparison with sysclk, then it is not able to cope
with all the instructions that it is supposed to do.

Signed-off-by: Sergey Matyukevich <geomatsi@gmail.com>