app v3: improve IR RC
Improve efficiency of IR remote control processing.
Key observations and changes:
- no need to poll: start polling when IR pin goes down
- enable decoding task (polling) in EXTI interrupt handler
  -- keep EXIT handler disabled for IR pin while polling
- disabling polling and re-enable EXTI IR pin interrupt
  after IR decode completion (error or success)
- grace period 1sec after successful decoding
- try again in ~25 msec on error

Room for improvement: think about more robust method
for recovery from errors. In fact error usually means
that we started decoding not from the start. So it makes
sense to wait for the next IR transmission. Current
solution is reasonably OK for RC buttons, but not yet
good enough for Dock beacons.

Signed-off-by: Sergey Matyukevich <geomatsi@gmail.com>