app v4: introduce decision sw layer
Move all the decision logic from app to 'brain' module.
Now app-v4 is a kind of HAL layer that provides events
to 'brain' module. Brain module in its turn is responsible
for decisions in the form of actions that are executed
by the HAL layer.

Implement simple 'brain' for debug purposes:
- debug all events
- beep on button press

Signed-off-by: Sergey Matyukevich <geomatsi@gmail.com>